Move the text functionality to a separate Text Coloring class.
