+++ patterns/adapter/flutter_adapter/client_app/app.dart
@@ -2,41 +2,20 @@ import 'dart:ui';
 
 import '../../../observer/app_observer/observer/event.dart';
 import '../adapter/classic_app.dart';
+import 'text_coloring.dart';
 
 class NextTextColorEvent extends Event {}
 
 class App extends ClassicAppBase {
-  final maxTextSize = 150;
-
-  var _textSize = 50;
-
-  int get textSize => _textSize;
-
-  set textSize(int newVal) {
-    if (newVal == _textSize || newVal > maxTextSize || newVal < 1) {
-      return;
-    }
-
-    _textSize = newVal;
-    repaint();
+  App() {
+    textColoring = TextColoring(this);
   }
 
-  var _textColor = Color(0xffd81b60);
-
-  Color get textColor => _textColor;
-
-  set textColor(Color newColor) {
-    if (_textColor == newColor) {
-      return;
-    }
-    print(newColor);
-    _textColor = newColor;
-    repaint();
-  }
+  late final TextColoring textColoring;
 
   @override
   void onWheel(double deltaX, double deltaY) {
-    textSize += deltaY ~/ 10;
+    textColoring.textSize += deltaY ~/ 10;
   }
 
   @override
@@ -46,28 +25,6 @@ class App extends ClassicAppBase {
 
   @override
   void onPaint(Canvas canvas, Size size) {
-    _paintTextCenter('Flutter Adapter', canvas, size);
-  }
-
-  void _paintTextCenter(String text, Canvas canvas, Size size) {
-    final builder = ParagraphBuilder(
-      ParagraphStyle(
-        textAlign: TextAlign.center,
-        fontSize: _textSize.toDouble(),
-      ),
-    )
-      ..pushStyle(
-        TextStyle(
-          fontFamily: 'Arial',
-          color: _textColor,
-        ),
-      )
-      ..addText(text);
-
-    final paragraph = builder.build()
-      ..layout(ParagraphConstraints(width: size.width));
-
-    final centerPos = Offset(0, (size.height - paragraph.height) / 2);
-    canvas.drawParagraph(paragraph, centerPos);
+    textColoring.paint('Flutter Adapter', canvas, size);
   }
 }
